+++ apps/web/app/(dashboard)/(admin)/admin/page.tsx
@@ -1,8 +1,8 @@
 import { type Metadata } from "next";
 import { METADATA } from "@/constants/metadata";
 import ChartsSection from "@/components/admin/charts-section";
-import UsersSection from "@/components/users-management/users-section";
 import Header from "@repo/ui/components/header";
+import UsersSection from "@/features/users-management/users-section";
 
 export const metadata: Metadata = {
   title: "Admin Dashboard",
